phase-2/backend/pyproject.toml
[project]
name = "phase2-backend"
version = "0.1.0"
description = "Phase II Todo App - FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.15.0",
    "bcrypt>=5.0.0",
    "dapr>=1.12.0",
    "email-validator>=2.0.0",
    "fastapi>=0.120,<0.125",
    "httpx>=0.28.1",
    "kafka-python>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.27",
    "slowapi>=0.1.9",
    "starlette>=0.49.1",
    "uvicorn[standard]>=0.34.0",
    "openai-agents>=0.0.2",
    "openai>=1.60.0",
    "mcp>=1.24.0",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pytest>=9.0.1",
]
